nampilin fups di list-usulan, complete edit&delete user pke alert

+++ resources/views/listUser.blade.php
@@ -18,6 +18,17 @@
     </h1>
     <a href="javascript:history.back()"><button class="btn btn-primary mx-5"><i class="fa fa-reply"></i>  Kembali</button></a>
     <div class="container my-4">
+        @if(session('alert'))
+            <div class="alert alert-success" role="alert">
+                <button type="button" class="close" data-dismiss="alert">x</button>
+                {{ session('alert')}}
+            </div>
+            @elseif(session('failed'))
+            <div class="alert alert-danger" role="alert">
+                <button type="button" class="close" data-dismiss="alert">x</button>
+                {{ session('failed')}}
+            </div>
+        @endif
         <table class="table table-bordered my-3">
             <thead>
                 <tr>
@@ -38,8 +49,8 @@
                     <td>{{$user->phone}}</td>
                     <td>{{$user->Bidang->name}}</td>
                     <td>
-                        <a href="#"><button class="btn btn-primary my-2 my-sm-0" type="submit"><i class="	fa fa-edit"></i></button></a>
-                        <a href="/Hapus/Data/User{{$user->id}}"><button class="btn btn-danger my-2 my-sm-0" type="submit"><i class="fa fa-trash"></i></button></a>
+                        <a href="/Edit/User/{{$user->id}}"><button class="btn btn-primary my-2 my-sm-0" type="submit"><i class="	fa fa-edit"></i></button></a>
+                        <a href="/Hapus/Data/User{{$user->id}}"><button class="btn btn-danger my-2 my-sm-0" type="submit" onclick="return confirm('Apakah anda yakin?');"><i class="fa fa-trash"></i></button></a>
                     </td>
                 </tr>
                 @endforeach
